requirements.txt
Flask==3.0.0
passlib==1.7.4
zxcvbn-python==4.4.24
python-dotenv==1.0.0
bcrypt==4.0.1
gunicorn==21.2.0
Werkzeug==3.0.1 